requirements.txt
mistralai
python-dotenv
streamlit
transformers
faiss-cpu